Refactor invitation code module directory structure
