+++ src/modules/invitationCode/entities/invitation-code.entity.ts
